Address review comments

+++ lib/matplotlib/hatch.py
@@ -170,7 +170,7 @@ class Stars(Shapes):
         path = Path.unit_regular_star(5)
         self.shape_vertices = path.vertices
         self.shape_codes = np.full(len(self.shape_vertices), Path.LINETO,
-                                   dtype='float64')
+                                   dtype='int32')
         self.shape_codes[0] = Path.MOVETO
         Shapes.__init__(self, hatch, density)
